+++ azalea-client/src/plugins/packet/mod.rs
@@ -45,7 +45,7 @@ impl Plugin for PacketPlugin {
             .add_message::<game::DeathEvent>()
             .add_message::<game::KeepAliveEvent>()
             .add_message::<game::ResourcePackEvent>()
-            .add_message::<game::InstanceLoadedEvent>()
+            .add_message::<game::WorldLoadedEvent>()
             .add_message::<login::ReceiveCustomQueryEvent>();
     }
 }