+++ azalea-client/src/plugins/tick_end.rs
@@ -4,7 +4,7 @@ use azalea_core::tick::GameTick;
 use azalea_entity::LocalEntity;
 use azalea_physics::PhysicsSystems;
 use azalea_protocol::packets::game::ServerboundClientTickEnd;
-use azalea_world::InstanceName;
+use azalea_world::WorldName;
 use bevy_app::{App, Plugin};
 use bevy_ecs::prelude::*;
 
@@ -27,7 +27,7 @@ impl Plugin for TickEndPlugin {
 }
 
 pub fn game_tick_packet(
-    query: Query<Entity, (With<LocalEntity>, With<InstanceName>)>,
+    query: Query<Entity, (With<LocalEntity>, With<WorldName>)>,
     mut commands: Commands,
 ) {
     for entity in query.iter() {