+++ azalea/src/swarm/events.rs
@@ -1,4 +1,4 @@
-use azalea_client::local_player::InstanceHolder;
+use azalea_client::local_player::WorldHolder;
 use azalea_core::entity_id::MinecraftEntityId;
 use bevy_app::{App, Plugin, Update};
 use bevy_ecs::prelude::*;
@@ -21,7 +21,7 @@ pub struct SwarmReadyEvent;
 struct IsSwarmReady(bool);
 
 fn check_ready(
-    query: Query<Option<&MinecraftEntityId>, With<InstanceHolder>>,
+    query: Query<Option<&MinecraftEntityId>, With<WorldHolder>>,
     mut is_swarm_ready: ResMut<IsSwarmReady>,
     mut ready_events: MessageWriter<SwarmReadyEvent>,
 ) {